Move all the layers and rewrite their imports to only use public API. #50.
