+++ src/connected/index.ts
@@ -1,8 +1,8 @@
 export * from './interfaces';
-export * from './layers';
 export * from './export-only/exportableActions';
 export * from './export-only/exportableSelectors';
 export * from './export-only/exportableState';
 export * from './dataUtils';
 export * from './model/constants';
 export { default as ChartProvider, Props as ChartProviderProps } from './ChartProvider';
+export { default as ConnectedResizeSentinelLayer } from './ConnectedResizeSentinelLayer';
