+++ src/core/index.ts
@@ -2,7 +2,6 @@ export * from './renderUtils';
 export * from './intervalUtils';
 export * from './componentUtils';
 export * from './decorators';
-export * from './layers';
 export * from './interfaces';
 export { default as Stack } from './Stack';
 export { default as propTypes } from './propTypes';
