+++ src/index.ts
@@ -1,3 +1,4 @@
 export * from './core';
 export * from './connected';
+export * from './layers';
 export * from './test-util';
