+++ src/layers/InteractionCaptureLayer.tsx
@@ -2,8 +2,11 @@ import * as React from 'react';
 import * as PureRender from 'pure-render-decorator';
 import * as d3Scale from 'd3-scale';
 
-import propTypes from '../propTypes';
-import { Interval, BooleanMouseEventHandler } from '../interfaces';
+import {
+  Interval,
+  BooleanMouseEventHandler,
+  propTypes
+} from '../core';
 
 const LEFT_MOUSE_BUTTON = 0;
