+++ src/layers/PollingResizingCanvasLayer.tsx
