+++ src/layers/connected/ConnectedXAxisLayer.tsx
@@ -1,9 +1,10 @@
 import * as React from 'react';
 import { connect } from 'react-redux';
 
-import { Interval, AxisSpec, XAxisLayer } from '../../core';
-import { ChartState } from '../model/state';
-import { selectXDomain } from '../model/selectors';
+import { Interval, AxisSpec } from '../../core';
+import { ChartProviderState, selectXDomain } from '../../connected';
+import XAxisLayer from '../XAxisLayer';
+
 
 export interface OwnProps extends AxisSpec {
   font?: string;
@@ -13,7 +14,7 @@ export interface ConnectedProps {
   xDomain: Interval;
 }
 
-function mapStateToProps(state: ChartState): ConnectedProps {
+function mapStateToProps(state: ChartProviderState): ConnectedProps {
   return {
     xDomain: selectXDomain(state)
   };
