+++ src/layers/connected/wrapDataLayerWithConnect.tsx
@@ -3,9 +3,7 @@ import * as PureRender from 'pure-render-decorator';
 import { connect } from 'react-redux';
 
 import { Interval, ScaleFunction, SeriesData} from '../../core';
-import { SeriesId } from '../interfaces';
-import { ChartState } from '../model/state';
-import { selectData, selectXDomain, selectYDomains } from '../model/selectors';
+import { ChartProviderState, SeriesId, selectData, selectXDomain, selectYDomains } from '../../connected';
 
 export interface SeriesIdProp {
   seriesId: SeriesId;
@@ -17,7 +15,7 @@ export interface WrappedDataLayerConnectedProps {
   yDomain: Interval;
 }
 
-function mapStateToProps(state: ChartState, ownProps: SeriesIdProp): WrappedDataLayerConnectedProps {
+function mapStateToProps(state: ChartProviderState, ownProps: SeriesIdProp): WrappedDataLayerConnectedProps {
   return {
     data: selectData(state)[ownProps.seriesId],
     xDomain: selectXDomain(state),
